+++ src/libWWW/HTTPS.c
@@ -113,7 +113,7 @@ extern char* HTAppVersion; /* Application version */
 */
 PRIVATE char* find_header(char* headers, const char* header_name) {
     char* p = headers;
-    int len = strlen(header_name);
+    int len = (int)strlen(header_name);
     
     while (*p) {
         /* Check if we're at start of line */
@@ -205,7 +205,7 @@ PUBLIC int HTLoadHTTPS ARGS4(const char*, arg,
     int foundContentLength = 0;
 
     if (delimSize == 0)
-        delimSize = strlen(delim);
+        delimSize = (int)strlen(delim);
 
     if (!arg)
         return -3;
